requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMongo==4.6.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
openai-whisper
groq>=0.9.0
httpx>=0.24.1
openai
torch
librosa
soundfile
tiktoken
